+++ cmd/substreams/service-stop.go
@@ -14,25 +14,23 @@ import (
 )
 
 func init() {
-	alphaCmd.AddCommand(sinkPauseCmd)
-	sinkPauseCmd.Flags().StringP("endpoint", "e", "http://localhost:8000", "specify the endpoint to connect to.")
-	sinkPauseCmd.Flags().Bool("strict", false, "Require deploymentID parameter to be set and complete")
+	serviceCmd.AddCommand(stopCmd)
 }
 
-var sinkPauseCmd = &cobra.Command{
-	Use:   "sink-pause [deployment-id]",
-	Short: "Pause a running substreams sink",
+var stopCmd = &cobra.Command{
+	Use:   "stop [deployment-id]",
+	Short: "Stop a running service",
 	Long: cli.Dedent(`
-        Sends an "Pause" request to a server. By default, it will talk to a local "substreams alpha sink-serve" instance.
-        It will pause a substreams and returns information about the change of status.
+        Sends an "Stop" request to a server. By default, it will talk to a local "substreams alpha service serve" instance.
+        It will stop a service and returns information about the change of status.
         If deploymentID is not set or is incomplete, the CLI will try to guess (unless --strict is set).
 		`),
-	RunE:         sinkPauseE,
+	RunE:         stopE,
 	Args:         cobra.RangeArgs(0, 1),
 	SilenceUsage: true,
 }
 
-func sinkPauseE(cmd *cobra.Command, args []string) error {
+func stopE(cmd *cobra.Command, args []string) error {
 	ctx := cmd.Context()
 
 	var id string
@@ -45,19 +43,22 @@ func sinkPauseE(cmd *cobra.Command, args []string) error {
 		if sflags.MustGetBool(cmd, "strict") {
 			return fmt.Errorf("invalid ID provided: %q and '--strict' is set", id)
 		}
-		matching, err := fuzzyMatchDeployment(ctx, id, cli, fuzzyMatchPreferredStatusOrder)
+		matching, err := fuzzyMatchDeployment(ctx, id, cli, cmd, fuzzyMatchPreferredStatusOrder)
 		if err != nil {
 			return err
 		}
 		id = matching.Id
 	}
 
-	req := &pbsinksvc.PauseRequest{
+	req := connect.NewRequest(&pbsinksvc.StopRequest{
 		DeploymentId: id,
+	})
+	if err := addHeaders(cmd, req); err != nil {
+		return err
 	}
 
-	fmt.Printf("Pausing... please wait\n")
-	resp, err := cli.Pause(ctx, connect.NewRequest(req))
+	fmt.Printf("Stopping... (shutting down services, please wait)\n")
+	resp, err := cli.Stop(ctx, req)
 	if err != nil {
 		return interceptConnectionError(err)
 	}